refactor: 主要な型を metricsModel に移動
