+++ src/feature/metric/Metrics.ts
@@ -1,5 +1,3 @@
 export default interface Metrics<T> {
   get metrics(): T;
 }
-
-export type MetricsScope = 'file' | 'function' | 'class' | 'method' | 'object';
